Bug fix and code update in entity model classes
Changed the 'pokéDex' in EggHatchEntityModel from Integer to int. Setting the type as Integer caused NPEs when reading batch files into pojos. 

Changed 'id' in all the entity models from Integer to long.

+++ src/main/java/com/thehumblefool/pokégoapi2/models/entities/EggHatchEntityModel.java
@@ -27,10 +27,10 @@ public class EggHatchEntityModel implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private Integer id;
+    private long id;
 
     @Column(name = "poké_dex")
-    private Integer pokéDex;
+    private int pokéDex;
 
     @Column(name = "pokémon")
     private String pokémon;
@@ -60,11 +60,11 @@ public class EggHatchEntityModel implements Serializable {
         this.id = id;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
